⚡:
Apply AI assistant changes to 2 files.
Ref: ⚡: ...
