+++ app/webhook-handlers/commands/ctx.ts
@@ -2,62 +2,12 @@ import { sendTelegramMessage, sendTelegramDocument } from "@/app/actions";
 import { sendComplexMessage, KeyboardButton } from "../actions/sendComplexMessage";
 import { logger } from "@/lib/logger";
 import { fetchRepoTree, fetchRepoContents } from "@/app/actions_github/actions";
-
+import { subcontexts } from "@/app/webhook-handlers/commands/content/subcontexts”
 
 const REPO_URL = process.env.NEXT_PUBLIC_REPO_URL || "https://github.com/salavey13/carTest"; // Fallback URL
 const MAX_FILE_SIZE = 10 * 1024 * 1024; // 10MB Limit (Telegram Doc Size)
 
-export const subcontexts = { // Export subcontexts list for direct access by name
-    "Core Application Logic": [
-        "next.config.mjs",
-        "package.json",
-        "tailwind.config.ts",
-        "tsconfig.json",
-        "types/database.types",
-        "app/layout.tsx",
-        "components/layout/",
-        "app/globals.css",
-        "contexts/AppContext.tsx",
-        "hooks/supabase.ts",
-        "lib/auth.ts",
-        "lib/logger.ts",
-        "lib/debugLogger.ts",
-    ],
-    "Arbitrage": [
-        "app/arbitrage-explained",
-        "app/arbitrage-live-scanner",
-        "app/arbitrage-notdummies",
-        "app/arbitrage-test-agent",
-        "components/arbitrage",
-        "app/elon", //contains "GodMode" features in subdirectories
-        "app/elon/testbase",
-        "supabase/migrations/market_data_1.sql",
-        "hooks/useTradeSimulator.ts",
-    ],
-    "Webhooks and Telegram Bot": [
-        "app/api/telegramWebhook/route.ts",
-        "app/webhook-handlers",
-        "app/actions.ts",
-        "telegram.d.ts",
-        "hooks/supabase.ts",
-    ],
-    "CyberFitness (Gamification)": [
-        "app/cyberfitness",
-        "hooks/cyberFitnessSupabase.ts",
-    ],
-    "PDF/TO Doc Processing": [
-        "app/topdf",
-        "app/todoc",
-    ],
-    "Tutorials/Examples": ["app/tutorials", "app/style-guide/page.tsx"],
-    "Vibe Content Renderer": [
-        "components/VibeContentRenderer.tsx",
-        "lib/iconNameMap.ts",
-    ],
-    "Lead Generation & CRM": [
-        "app/leads",
-    ],
-};
+
 
 async function fetchFilesForSubcontext(repoUrl: string, tree: any[], filesPaths: string[]): Promise<{ success: boolean; files?: any[]; error?: string }> {
     if (!repoUrl) return { success: false, error: "Repo URL is missing." };
@@ -99,6 +49,7 @@ export async function ctxCommand(chatId: number, userId: number) {
     logger.info(`[Ctx Command V2] User ${userId} triggered the /ctx command.`);
 
     if (!REPO_URL) {
+
         logger.error("[Ctx Command] REPO_URL is not defined in environment variables.");
         await sendTelegramMessage(
             "🚨 Ошибка: URL репозитория не настроен на сервере.",
